requirements.txt
# 公众号写作助手 - Python 依赖
# WeChat Article Wizard - Dependencies
# 最低要求 Python 3.10，推荐 Python 3.11 或 3.12

# HTTP 客户端 (用于API调用)
# 注意：Python 3.14 用户请使用 httpx>=0.28.0
httpx>=0.28.0
aiohttp>=3.9.0

# 环境配置
python-dotenv>=1.0.0

# Web 框架 (可选，用于本地服务)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
aiosqlite>=0.19.0

# 工具库
python-multipart>=0.0.6
